requirements.txt
websockets>=10.0
python-binance>=1.0.15
pandas>=1.3.0
numpy>=1.21.0